code: Заменил тип переменных a и b на double

+++ ConsoleApplication1/ConsoleApplication1.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main()
 {
     std::cout << "Hello World!\n";
-    int a, b,sum,raz,proizv,ost;
+    double a, b,sum,raz,proizv,ost;
     cin >> a >> b;
     sum = a + b;
     raz = a - b;
